+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add flag on project templates to mark as active or inactive [#626](https://github.com/PublicMapping/districtbuilder/pull/626)
 - Display featured projects on Organization page [#633](https://github.com/PublicMapping/districtbuilder/pull/633)
 - Add import plan page [#639](https://github.com/PublicMapping/districtbuilder/pull/639)
+- Add project evaluate view for compactness [#646](https://github.com/PublicMapping/districtbuilder/pull/646)
 
 ### Changed